requirements/openmmlab_windows.txt
mmengine==0.10.7
mmsegmentation==1.2.2
mmdetection==3.3.0
openmim==0.3.9
ftfy==6.3.1
numpy==1.26.4
opencv-python==4.10.0.84
